+++ app/src/main/java/com/example/bittrack/ui/home/HomeEvent.kt
@@ -1,6 +1,8 @@
 package com.example.bittrack.ui.home
 
+import java.math.BigDecimal
+
 sealed class HomeEvent {
     data object GetBtcRate : HomeEvent()
-    data class AddIncomeTransaction(val amount: Double) : HomeEvent()
+    data class AddIncomeTransaction(val amount: BigDecimal) : HomeEvent()
 }
